acutis_old/pyproject.toml
[project]
name = "hesed-api"
version = "0.1.0"
description = "Backend do projeto Acutis"
readme = "README.md"
requires-python = ">=3.11.9"
dependencies = [
    "aiosignal==1.3.1 ; python_full_version >= '3.7'",
    "alembic==1.11.1 ; python_full_version >= '3.7'",
    "apscheduler==3.10.1 ; python_full_version >= '3.6'",
    "argon2-cffi==23.1.0 ; python_full_version >= '3.7'",
    "argon2-cffi-bindings==21.2.0 ; python_full_version >= '3.6'",
    "asttokens==3.0.0 ; python_full_version >= '3.8'",
    "async-timeout==4.0.2 ; python_full_version >= '3.6'",
    "attrs==23.1.0 ; python_full_version >= '3.7'",
    "azure-core==1.29.3 ; python_full_version >= '3.7'",
    "azure-storage-blob==12.17.0 ; python_full_version >= '3.7'",
    "blinker==1.6.2 ; python_full_version >= '3.7'",
    "boto3==1.34.149 ; python_full_version >= '3.8'",
    "botocore==1.34.149 ; python_full_version >= '3.8'",
    "certifi==2023.5.7 ; python_full_version >= '3.6'",
    "cffi==1.15.1",
    "cfgv==3.4.0 ; python_full_version >= '3.8'",
    "chardet==3.0.4",
    "charset-normalizer==3.2.0 ; python_full_version >= '3.7'",
    "click==8.1.3 ; python_full_version >= '3.7'",
    "colorama==0.4.6 ; python_full_version >= '2.7'",
    "coverage[toml]==7.6.9 ; python_full_version >= '3.9'",
    "cryptography==41.0.3 ; python_full_version >= '3.7'",
    "deprecated==1.2.14 ; python_full_version >= '2.7'",
    "distlib==0.3.9",
    "dnspython==2.3.0 ; python_full_version >= '3.7' and python_full_version < '4.0'",
    "email-validator==2.0.0.post2 ; python_full_version >= '3.7'",
    "et-xmlfile==1.1.0 ; python_full_version >= '3.6'",
    "executing==2.1.0 ; python_full_version >= '3.8'",
    "factory-boy==3.3.1 ; python_full_version >= '3.8'",
    "faker==33.3.0 ; python_full_version >= '3.8'",
    "filelock==3.16.1 ; python_full_version >= '3.8'",
    "flask==2.3.2 ; python_full_version >= '3.8'",
    "flask-apscheduler==1.12.4",
    "flask-cors==3.0.10",
    "flask-jwt-extended==4.5.2 ; python_full_version >= '3.7' and python_full_version < '4.0'",
    "flask-limiter==3.6.0 ; python_full_version >= '3.8'",
    "flask-mail==0.9.1",
    "flask-migrate==4.0.4 ; python_full_version >= '3.6'",
    "flask-sqlalchemy==3.0.3 ; python_full_version >= '3.7'",
    "frozenlist==1.4.0 ; python_full_version >= '3.8'",
    "googlemaps==4.10.0 ; python_full_version >= '3.5'",
    "googletrans==4.0.0rc1 ; python_full_version >= '3.6'",
    "greenlet==2.0.2 ; python_full_version >= '2.7'",
    "gunicorn==20.1.0 ; python_full_version >= '3.5'",
    "h11==0.9.0",
    "h2==3.2.0",
    "hpack==3.0.0",
    "hstspreload==2025.1.1 ; python_full_version >= '3.6'",
    "httpcore==0.9.1 ; python_full_version >= '3.6'",
    "httpx==0.13.3 ; python_full_version >= '3.6'",
    "hyperframe==5.2.0",
    "icecream==2.1.3",
    "identify==2.6.2 ; python_full_version >= '3.9'",
    "idna==2.10 ; python_full_version >= '2.7'",
    "importlib-resources==6.4.0 ; python_full_version >= '3.8'",
    "iniconfig==2.0.0 ; python_full_version >= '3.7'",
    "isodate==0.6.1",
    "itsdangerous==2.1.2 ; python_full_version >= '3.7'",
    "jinja2==3.1.2 ; python_full_version >= '3.7'",
    "jmespath==1.0.1 ; python_full_version >= '3.7'",
    "libretranslatepy==2.1.1",
    "limits==3.11.0 ; python_full_version >= '3.8'",
    "lxml==5.2.1 ; python_full_version >= '3.6'",
    "mako==1.2.4 ; python_full_version >= '3.7'",
    "markdown-it-py==3.0.0 ; python_full_version >= '3.8'",
    "markupsafe==2.1.2 ; python_full_version >= '3.7'",
    "mdurl==0.1.2 ; python_full_version >= '3.7'",
    "mercadopago==2.2.1 ; python_full_version >= '3'",
    "minio==7.2.7",
    "multidict==6.0.4 ; python_full_version >= '3.7'",
    "nodeenv==1.9.1 ; python_full_version >= '2.7'",
    "numpy==2.1.1 ; python_full_version >= '3.10'",
    "openpyxl==3.1.5 ; python_full_version >= '3.8'",
    "ordered-set==4.1.0 ; python_full_version >= '3.7'",
    "packaging==24.0 ; python_full_version >= '3.7'",
    "pandas==2.2.2 ; python_full_version >= '3.9'",
    "pillow==10.3.0 ; python_full_version >= '3.8'",
    "platformdirs==4.3.6 ; python_full_version >= '3.8'",
    "pluggy==1.5.0 ; python_full_version >= '3.8'",
    "prometheus-client>=0.21.1",
    "pycparser==2.21",
    "pycryptodome==3.20.0 ; python_full_version >= '2.7'",
    "pydantic==1.10.8 ; python_full_version >= '3.7'",
    "pygments==2.17.2 ; python_full_version >= '3.7'",
    "pyjwt==2.7.0 ; python_full_version >= '3.7'",
    "pyodbc==4.0.39 ; python_full_version >= '2.7'",
    "pytest==8.3.4 ; python_full_version >= '3.8'",
    "pytest-cov==6.0.0 ; python_full_version >= '3.9'",
    "python-dateutil==2.8.2 ; python_full_version >= '2.7'",
    "python-dotenv==1.0.0 ; python_full_version >= '3.8'",
    "python-http-client==3.3.7 ; python_full_version >= '2.7'",
    "pytz==2023.3",
    "pyyaml==6.0.2 ; python_full_version >= '3.8'",
    "redis==5.2.1 ; python_full_version >= '3.8'",
    "requests==2.31.0 ; python_full_version >= '3.7'",
    "rfc3986==1.5.0",
    "rich==13.7.1 ; python_full_version >= '3.7'",
    "s3transfer==0.10.2 ; python_full_version >= '3.8'",
    "sendgrid==6.10.0 ; python_full_version >= '2.7'",
    "sentry-sdk[flask]==2.20.0 ; python_full_version >= '3.6'",
    "setuptools==75.8.0 ; python_full_version >= '3.9'",
    "six==1.16.0 ; python_full_version >= '2.7'",
    "sniffio==1.3.1 ; python_full_version >= '3.7'",
    "spectree==1.1.2 ; python_full_version >= '3.6'",
    "sqlalchemy==2.0.15 ; python_full_version >= '3.7'",
    "starkbank-ecdsa==2.2.0",
    "translate==3.6.1",
    "typing-extensions==4.6.2 ; python_full_version >= '3.7'",
    "tzdata==2023.3 ; python_full_version >= '2'",
    "tzlocal==5.0.1 ; python_full_version >= '3.7'",
    "urllib3==2.0.3 ; python_full_version >= '3.7'",
    "validate-docbr==1.10.0",
    "virtualenv==20.27.1 ; python_full_version >= '3.8'",
    "werkzeug==2.3.4 ; python_full_version >= '3.8'",
    "wrapt==1.16.0 ; python_full_version >= '3.6'",
    "xlsxwriter==3.2.0 ; python_full_version >= '3.6'",
    "xmltodict==0.13.0 ; python_full_version >= '3.4'",
    "yarl==1.9.2 ; python_full_version >= '3.7'",
]

[dependency-groups]
dev = [
    "pre-commit>=4.1.0",
]
